requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
numpy==1.26.4
pandas==2.2.3
xarray==2024.11.0
h5py==3.12.1
h5netcdf==1.4.1
netCDF4==1.7.2
scipy==1.15.2
torch==2.5.1
scikit-learn==1.6.1
pennylane==0.40.0
autoray==0.7.2
huggingface_hub[hf_xet]==0.30.1
python-dotenv==1.0.1
pydantic==2.10.6
pyyaml==6.0.2
requests==2.32.3
orjson==3.10.15
aiofiles==24.1.0
httpx==0.28.1
rich==13.9.4
typer==0.15.1
tqdm==4.67.1
git+https://github.com/NASA-IMPACT/Prithvi-WxC.git@main
